chore: modularize and refactor
